refactor: flatten structure
remove pkg folder and refactor imports

using pkg folder seems out of place in a library with no entry point
