+++ examples/cli/main.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/evilmonkeyinc/golang-cli/pkg/commands"
-	"github.com/evilmonkeyinc/golang-cli/pkg/middleware"
-	"github.com/evilmonkeyinc/golang-cli/pkg/shell"
+	"github.com/evilmonkeyinc/golang-cli/commands"
+	"github.com/evilmonkeyinc/golang-cli/middleware"
+	"github.com/evilmonkeyinc/golang-cli/shell"
 )
 
 func main() {
