+++ shell/flaghandler.go
