+++ shell/flaghandler_test.go
