+++ shell/flagset_test.go
