+++ shell/handler_test.go
