+++ shell/middleware_test.go
