+++ shell/request_test.go
