+++ shell/shell_test.go
