+++ shell/writer.go
