+++ shell/writer_test.go
